Add files via upload
This Tic-Tac-Toe game project is a simple, interactive implementation of the classic 3x3 grid game, where two players take turns to mark spaces with 'X' and 'O' until one player aligns three of their marks horizontally, vertically, or diagonally. The project is built using javascript, css, and html, and features a user-friendly interface that allows players to easily select their moves. It also includes functionality to check for win conditions, declare the winner, and restart the game after a completed round. This project serves as a great way to practice fundamental programming concepts such as conditionals, loops, and event handling.